% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{aril}
\alias{aril}
\title{Average Relative Interval Length (ARIL)}
\usage{
aril(Y, X, alpha = 0.5)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{predictions matrix [N x D]}

\item{alpha}{specifies 1 - alpha confidence level [0 < alpha < 1]}
}
\value{
average relative interval length [0 < scalar < 1]
}
\description{
This function calculates the average relative interval length (ARIL)
given a set of predictions 'X', the target 'Y', and a confidence level
of 1 - alpha.
}
\references{
Dogulu, N., P. Lopez Lopez, D.P. Solomatine, A.H. Weerts and
D.L. Shreshta (2015), Estimation of predictive hydrologic uncertainty
using the quantile regression and UNEEC methods and their comparison
on constrasting catchments, Hydrol. Earth. Syst. Sci., 19, 3181-3201.
}
