% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{avg.asym.deg.S}
\alias{avg.asym.deg.S}
\title{S index of average asymmetry degree}
\usage{
avg.asym.deg.S(Y, X, alpha = 0.5)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{predictions matrix [N x D]}

\item{alpha}{specifies 1 - alpha confidence level [0 < alpha < 1], Default: 0.5}
}
\value{
S index of average asymmetry degree [scalar]
}
\description{
This function calculates the S index of average asymmetry degree given a
set of predictions 'X', the target 'Y', and a confidence level of
1 - alpha.
}
\references{
Xiong, L., N. Wan, X. Wei, and K. O'Connor (2009), Indices
for assessing the prediction bounds of hydrological models
and application by generalised likelihood uncertainty
estimation, Hydrological Sciences Journal, 54:5, 852-871.
}
