% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bindYX}
\alias{bindYX}
\title{Column Bind Target and Input Variables}
\usage{
bindYX(y, x)
}
\arguments{
\item{y}{target time series [N x 1]}

\item{x}{input time series [N x D]}
}
\value{
resulting matrix [N x D+1]
}
\description{
This utility function column binds target and input variables together in a data frame
and removes 'NA' values.
}
