% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bindYXZ}
\alias{bindYXZ}
\title{Column Bind Target, Input and Auxiliary Variables}
\usage{
bindYXZ(y, x, z)
}
\arguments{
\item{y}{target time series [N x 1]}

\item{x}{input time series [N x D]}

\item{z}{auxiliary variables [N x M]}
}
\value{
resulting matrix [N x D+M+1]
}
\description{
This function column binds target, input variables, and auxiliary variables together
in a data frame and removes 'NA' values.
}
