% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{calibrateDDM}
\alias{calibrateDDM}
\title{Calibrate (train) a data-driven model (DDM)}
\usage{
calibrateDDM(yc, xc, ddm, ddm_param)
}
\arguments{
\item{yc}{calibration target [N x 1]}

\item{xc}{calibration inputs [N x D]}

\item{ddm}{data-driven model [string: 'spov','knnr','grnn','rrf']}

\item{ddm_param}{data-driven model hyper-parameter(s)
\itemize{
\item{"spov"}{ddm_param[1] = model order [1 < integer <= 3]}
\item{"knnr"}{ddm_param[1] = no. of nearest neighbours [1 < integer < sample size-1]}
\item{"grnn"}{no ddm_param as kernel bandwidth determined automatically}
\item{"rrf"}{ddm_param[1] = ntrees [(1,500)]}
}}
}
\value{
integer, list, etc. that contains calibrated model structure
}
\description{
This function calibrates a data-driven model (DDM) for a given input-
output dataset.
}
\references{
Quilty, J., J. Adamowski, B. Khalil, and M. Rathinasamy (2016), Bootstrap rank-
ordered conditional mutual information (broCMI): A nonlinear input variable
selection method for water resources modeling, Water Resour. Res., 52,
doi:10.1002/2015WR016959.

Quilty, J., and J. Adamowski (2018), Addressing the incorrect usage of wavelet-
based hydrological and water resources forecasting models for real-world
applications with best practices and a new forecasting framework, J. Hydrol.,
doi:10.1016/j.jhydrol.2018.05.003.
}
