% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{cpp}
\alias{cpp}
\title{Coverage Probability Plot (CPP)}
\usage{
cpp(Y, X)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{predictions matrix [N x D]}
}
\value{
list containing:
\itemize{
\item{"fq"}{forecast quantiles(0,0.1,...,1)}
\item{"tq"}{theoretical quantiles(0,0.1,...,1)}
\item{"cpp_mse"}{mean square error between forecast and theoretical quantiles}
}
}
\description{
This function calculates the coverage probability plot (CPP)
given a set of predictions 'X' and the target 'Y'.
}
\references{
F. Laio and S. Tamea (2007), Verification tools for probabilistic
forecasts of continuous hydrological variables,
Hydrol. Earth Syst. Sci., 11, pp. 1267-1277.

B. Renard, D. Kavetski, G. Kuczera, M. Thyer, and S.W. Franks (2009), Understanding
predictive uncertainty in hydrologic modeling: the challenge of identifying input
and structural errors. Water. Resour. Res., 46, p. W05521.
}
