% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{de_scale_ab}
\alias{de_scale_ab}
\title{De-scale Columns of a Matrix Based on Another Matrix}
\usage{
de_scale_ab(X1, X2, a, b, ...)
}
\arguments{
\item{X1}{input matrix whose properties are used in de-scaling [N x D]}

\item{X2}{input matrix to de-scale based on properties of X1 [M x D]}

\item{a}{lower limit value for scaled data [scalar]}

\item{b}{upper limit value for scaled data [scalar]}

\item{...}{further arguments to the \code{min()} and \code{max()} functions}
}
\value{
de-scaled input matrix X2 [M x D]
}
\description{
This function de-scales an input matrix 'X2' from between values 'a' and 'b' with
the option to disclude NA values from the scaling (see example below) according to
the content (max and min values) of scaling matrix 'X1'. In other words, this function
undoes what \code{scale_ab} does.
}
\examples{
\dontrun{
if(interactive()){
 X1 = matrix(runif(16),4,4)
 X2 = matrix(rnorm(16),4,4)
 X1[2:3,3:4] = NA
 X3 = de_scale_ab(X1,X2,-1,1,na.rm=TRUE)
 }
}
}
\references{
https://stackoverflow.com/questions/5468280/scale-a-series-between-two-points
}
