% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{grnn_estimate}
\alias{grnn_estimate}
\title{Calibrate a Generalized Regression Neural Newtork (GRNN)}
\usage{
grnn_estimate(Y, X, Yv, Xv)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{input matrix (explanatory variables) [N x D]}

\item{Yv}{validation target vector [Nv x 1]}

\item{Xv}{validation input matrix for generating validation predictions [Nv x D]}
}
\value{
optimized kernel bandwith [scalar]
}
\description{
This function calibrates a Generalized Regression Neural Network
model's kernel bandwidth 'sigma' based on the radial basis
function (RBF, also known as Gaussian) kernel using the 'optimize'
function from the 'stats' package for a set of calibration data
pairs (Y,X) and a set of validation data pairs (Yv,Xv).
}
\details{
Note that the number of distance calculations required for estimating the RBF
kernels (parameters) in the GRNN model grows O(N^2).
}
\references{
D.F. Specht (1991), A general regression neural network, IEEE Transactions
on Neural Networks, 2, 6, pp. 568-576, DOI: 10.1109/72.97934.

T.I. Harrold, A. Sharma, and S. Sheather (2001), Selection of a kernel
bandwidth for measuring dependence in hydrologic time series using the
mutual information criterion, Stoch. Environ. Res. Risk Assess., 15, pp.
310-324.
}
