% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{grnn_predict}
\alias{grnn_predict}
\title{Predict using a calibrated Generalized Regression Neural Network (GRNN) model}
\usage{
grnn_predict(
  Y,
  X,
  Xtest,
  bw = ((4/(nrow(X) + 2))^(1/(nrow(X) + 4))) * N^(-1/(ncol(X) + 4))
)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{input matrix (explanatory variables) [N x D]}

\item{Xtest}{test input matrix for generating predictions [Ntest x D]}

\item{bw}{RBF kernel bandwidth [scalar], Default: ((4/(nrow(X) + 2))^(1/(nrow(X) + 4))) * N^(-1/(ncol(X) + 4))}
}
\value{
prediction for input matrix Xtest [Ntest x 1]
}
\description{
This function calculates predictions for a set of model inputs 'Xtest'
using a Generalized Regression Neural Network model for a given kernel
bandwidth 'sigma' and using the radial basis function (RBF, also known
as Gaussian) kernel.
}
\details{
Default value of the \code{bw} input argument set to the Gaussian Reference Rule (Harrold et al., 2001),
see below in the references.
Note that the number of distance calculations required for estimating the RBF
kernels (parameters) in the GRNN model grows O(N^2).
}
\references{
D.F. Specht (1991), A general regression neural network, IEEE Transactions
on Neural Networks, 2, 6, pp. 568-576, DOI: 10.1109/72.97934.

T.I. Harrold, A. Sharma, and S. Sheather (2001), Selection of a kernel
bandwidth for measuring dependence in hydrologic time series using the
mutual information criterion, Stoch. Environ. Res. Risk Assess., 15, pp.
310-324.
}
