% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{lme}
\alias{lme}
\title{Liu-Mean Efficiency (LME)}
\usage{
lme(Y, X)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{prediction vector  [N x 1]}
}
\value{
Liu-Mean Efficiency (LME) [-inf < scalar <= 1]
}
\description{
This function calculates the Liu-Mean Efficiency (LME), a modified version of
the Nash-Sutcliffe (NSE) and Kling-Gupta Efficiency (KGE), given a vector of
predictions 'X' and targets 'Y'.
}
\references{
Liu, D. (2020), A rational performance criterion for hydrological model, J. Hydrol.,
580, 125488.
}
