% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{meanIS}
\alias{meanIS}
\title{Mean Interval Score (meanIS)}
\usage{
meanIS(Y, X, alpha = 0.5)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{predictions matrix [N x D]}

\item{alpha}{specifies 1 - alpha confidence level [0 < alpha < 1], Default: 0.5}
}
\value{
mean interval score [scalar]
}
\description{
This function calculates the mean interval score (meanIS)
given a set of predictions 'X', the target 'Y', and a
confidence level of 1 - alpha.
}
\references{
T. Gneiting and  A.E. Raferty (2007), Strictly proper score rules,
prediction, and estimation, Journal of the American Statistical
Association, 102, pp. 359-378.
F. Bourgin et al. (2015), Transferring global uncertainty estimates
from gauged to ungauged catchments, Hydrol. Earth Syst. Sci., 19, pp.
2535-2546.
C. Wan et al. (2014), Probabiistic forecasting of wind power
generation using extreme learning machine, IEEE Transactions on Power
Systems, 29, pp. 1033-1044.
}
