% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partitionInds}
\alias{partitionInds}
\title{Get train, validation and test set indices}
\usage{
partitionInds(N, nval, ntst)
}
\arguments{
\item{N}{number of time series records [scalar]}

\item{nval}{number of validation records [scalar]}

\item{ntst}{number of test records [scalar]}
}
\value{
list that holds training, validation, and test indices
}
\description{
This function provides train, validation and test set indices in a sequential fashion.
The user must provide the desired validation set size \code{nval} and test set size \code{ntst}.
}
