% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{picp}
\alias{picp}
\title{Prediction Interval Coverage Probability (PICP)}
\usage{
picp(Y, X, alpha = 0.5)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{predictions matrix [N x D]}

\item{alpha}{specifies 1 - alpha confidence level [0 < alpha < 1], Default: 0.5}
}
\value{
prediction interval coverage probability [0 < scalar < 1]
}
\description{
This function calculates the prediction interval coverage probability
(PICP) given a set of predictions 'X', the target 'Y', and a
confidence level of 1 - alpha.
}
\references{
Shrestha, D.L., N. Kayastha, D. Solomatine, and R. Price (2014),
Encapsulation of parametric uncertainty statistics by various
predictive machine leanring models: MLUE method, J. Hydroinformatics,
16, 95-113.
}
