% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wddff_utils.R
\name{prepareData}
\alias{prepareData}
\title{CHECK INPUTS AND SET COLUMN NAMES FOR DIFFERENT VARIABLES
perform time lag of target (and potentially standard inputs; note: auxulliary
inputs are not time-lagged)
if needed perform wavelet decomposition on target and inputs (but not auxilliary
inputs...)
output required data}
\usage{
prepareData(
  y,
  x = NULL,
  z = NULL,
  leadtime,
  lag_Y,
  lag_X,
  wfm = "none",
  wt = "at",
  wavelet = "haar",
  decomp_level = 1
)
}
\description{
CHECK INPUTS AND SET COLUMN NAMES FOR DIFFERENT VARIABLES
perform time lag of target (and potentially standard inputs; note: auxulliary
inputs are not time-lagged)
if needed perform wavelet decomposition on target and inputs (but not auxilliary
inputs...)
output required data
}
