% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{rnp}
\alias{rnp}
\title{Non-parametric Kling-Gupta Efficiency}
\usage{
rnp(Y, X)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{prediction vector [N x 1]}
}
\value{
non-parametric Kling-Gupta Efficiency (RNP) [-inf < scalar <= 1]
}
\description{
This function calculates the non-parametric Kling-Gupta Efficiency (RNP), a
modified version of the Kling-Gupta Efficiency (KGE), given a vector of
predictions 'X' and targets 'Y'.
}
\references{
Pool, S., Vis, M., Seibert, J. (2018), Evaluating model performance: towards a
non-parametric variant of the Kling-Gupta effficiency, Hydrological Sci. J.,
63 (13-14), 1941-1953.
}
