% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_ab}
\alias{scale_ab}
\title{Scale Columns of a Matrix Based on Another Matrix}
\usage{
scale_ab(X1, X2, a, b, ...)
}
\arguments{
\item{X1}{input matrix whose properties are used in scaling [N x D]}

\item{X2}{input matrix to scaled based on properties of X1 [M x D]}

\item{a}{lower limit value for scaled data [scalar]}

\item{b}{upper limit value for scaled data [scalar]}

\item{...}{further arguments to the \code{min()} and \code{max()} functions}
}
\value{
scaled input matrix X2 [M x D]
}
\description{
This function scales an input matrix 'X2' between values 'a' and 'b' with the
option to exclude NA values from the scaling (see example below) according to
the content (max and min values) of scaling matrix 'X1'.
}
\examples{
\dontrun{
if(interactive()){
 X1 = matrix(runif(16),4,4) # square matrix
 X1[2:3,3:4] = NA           # inject some NA's into X1
 X2 = matrix(rnorm(16),4,4) # another square matrix

 # Compute statistics with X1, apply scaling to X2
 X3 = scale_ab(X1,X2,-1,1,na.rm=TRUE)

 # NOTE: Typically, X1 will be the training set input features,
 #       X2 will be the validation/test set.
 }
}
}
\references{
https://stackoverflow.com/questions/5468280/scale-a-series-between-two-points
}
