% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{spoV_estimate}
\alias{spoV_estimate}
\title{Calibrate (train) a sparse p-th order Volterra series model}
\usage{
spoV_estimate(Y, X, p)
}
\arguments{
\item{Y}{target vector (response variable) [N x 1]}

\item{X}{input matrix (explanatory variables) [N x D]}

\item{p}{order (polynomial degree) of Volterra model [scalar]}
}
\value{
p-th order kernels of Volterra model [Q x 1]
}
\description{
This function calculates a  sparse p-th order multiple input single output
(MISO) Volterra series model for given input-output data pairs X and Y,
using the Moore-Penrose generalized inverse to estimate the different
triangular kernels (see Eq. 8 in Wu and Kareem [2014]).
}
\details{
NOTE: the number of kernels (parameters) in the Volterra series model grows exponentially
based on p (the model order).
}
\examples{
\dontrun{
if(interactive()){
# Usage:
p=2
X=matrix(runif(1000),500,20)
Y=matrix(runif(500),500,1)
H = spoV_estimate(Y,X,2) # sparse second-order Volterra series kernels
 }
}
}
\references{
T. Wu and A. Kareem (2014), Simulation of nonlinear bridge aerodynamics: A sparse
third-order Volterra model, Journal of Sound and Vibration, 333, 1, pp. 178-188
https://doi.org/10.1016/j.jsv.2013.09.003.

https://stackoverflow.com/questions/49538911/r-given-a-matrix-and-a-power-produce-multiple-matrices-containing-all-unique
}
