% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{spoV_predict}
\alias{spoV_predict}
\title{Predict using a calibrated sparse p-th order Volterra series model}
\usage{
spoV_predict(X, H, p)
}
\arguments{
\item{X}{input matrix (explanatory variables) [N x D] (must be a matrix (i.e., ?as.matrix)}

\item{H}{p-th order kernels of Volterra model [Q x 1]}

\item{p}{order (polynomial degree) of Volterra model [scalar]}
}
\value{
prediction [N x 1]
}
\description{
This function generates predictions for a sparse p-th order multiple input single
output (MISO) Volterra series model for given input X and parameters H.
}
\examples{
\dontrun{
if(interactive()){
# Usage:
p=2
X=matrix(runif(1000),500,20)
Y=matrix(runif(500),500,1)
H = spoV_estimate(Y,X,2); # sparse second-order Volterra series kernels
P = spoV_predict(X,H,p); # predictions
 }
}
}
\references{
T. Wu and A. Kareem (2014), Simulation of nonlinear bridge aerodynamics: A sparse
third-order Volterra model, Journal of Sound and Vibration, 333, 1, pp. 178-188
https://doi.org/10.1016/j.jsv.2013.09.003.
}
