% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Standardize the Columns of a Matrix Based on Another Matrix}
\usage{
standardize(X1, X2, ...)
}
\arguments{
\item{X1}{input matrix whose properties are used in standardization [N x D]}

\item{X2}{input matrix to standardize based on properties of X1 [M x D]}

\item{...}{further arguments to \code{mean()} and \code{std()}}
}
\value{
standardized input matrix X2 [M x D]
}
\description{
This function standardizes an input matrix 'X2' with the option to exclude
NA values from the scaling (see example below) according to the content
(mean and standard deviation values) of standardizing matrix 'X1'.
}
\examples{
\dontrun{
if(interactive()){
 X1 = matrix(runif(16),4,4)
 X2 = matrix(rnorm(16),4,4)
 X1[2:3,3:4] = NA
 X3 = standardize(X1,X2,na.rm=TRUE)
 }
}
}
\references{
https://stackoverflow.com/questions/5468280/scale-a-series-between-two-points
}
