% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wddff.R
\name{upd.wddff}
\alias{upd.wddff}
\title{Update WDDFF With New Data}
\usage{
upd.wddff(mdl_rslt, y, x = NULL, z = NULL, light = TRUE, savefile = FALSE)
}
\arguments{
\item{mdl_rslt}{list that contains all essential inputs from 'wddff'}

\item{y}{new target time series [N x 1]}

\item{x}{new input time series [N x D]}

\item{z}{new auxilliary input time series [N x C]}

\item{light}{store all function inputs or neglect 'y', 'x', and 'z' [binary]}

\item{savefile}{save file using function name and current Sys.time() [binary]}
}
\value{
list that contains:
\itemize{
\item mdl
\item predictions
\item target
\item residual
\item n.inds
\item indc
\item indv
\item indt
\item ivs
\item ddm
\item perf_c
\item perf_v
\item perf_t
}
}
\description{
Update WDDFF With New Data
}
\references{
This function takes in new time series measurements for the target and
input time series (standard and/or auxilliary), performs time-lagging,
(wavelet-decomposition,) and generates predictions given an exisiting
Wavelet Data-Driven Forecasting Framework (WDDFF) model.

Quilty, J., J. Adamowski, B. Khalil, and M. Rathinasamy (2016), Bootstrap rank-
ordered conditional mutual information (broCMI): A nonlinear input variable
selection method for water resources modeling, Water Resour. Res., 52,
doi:10.1002/2015WR016959.

Quilty, J., and J. Adamowski (2018), Addressing the incorrect usage of wavelet-
based hydrological and water resources forecasting models for real-world
applications with best practices and a new forecasting framework, J. Hydrol.,
doi:10.1016/j.jhydrol.2018.05.003.
}
