% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wddff_utils.R
\name{waveletDecomp}
\alias{waveletDecomp}
\title{Wavelet Decomposition}
\usage{
waveletDecomp(y, wt = "modwt", wavelet = "haar", decomp_level = 1)
}
\arguments{
\item{y}{time series [N x 1]}

\item{wt}{type of wavelet transform [string; 'at','modwt'], Default: 'modwt'}

\item{wavelet}{scaling filter name [string], Default: 'haar'}

\item{decomp_level}{decomposition level [1 < integer << N/2], Default: 1}
}
\value{
wavelet and scaling coefficients [N x D x J+1] (wavelet coefs in first J columns, scaling coefs in last column)
}
\description{
This function performs wavelet decomposition using either the maximal
overlap discrete wavelet transform (MODWT) or the à trous (AT) algorithm.
}
\references{
Quilty, J., and J. Adamowski (2018), Addressing the incorrect usage of wavelet-
based hydrological and water resources forecasting models for real-world
applications with best practices and a new forecasting framework, J. Hydrol.,
doi:10.1016/j.jhydrol.2018.05.003.
}
