% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wddff.R
\name{wddff}
\alias{wddff}
\title{Wavelet Data-Driven Forecasting Framework (WDDFF)}
\usage{
wddff(
  y,
  x = NULL,
  z = NULL,
  leadtime = 1,
  lag_Y = 1,
  lag_X = 1,
  nval = 1,
  ntst = 1,
  wfm = "none",
  wt = "at",
  wavelet = "haar",
  decomp_level = 1,
  max_decomp_level = 0,
  max_wavelet_length = 0,
  ivsm = "none",
  ivs_param = c(0.1, 128),
  ddm = "spov",
  ddm_param = 1,
  scale_inputs = FALSE,
  scale_target = FALSE,
  cutoff0 = FALSE,
  light = TRUE,
  savefile = FALSE
)
}
\arguments{
\item{y}{target time series [N x 1]}

\item{x}{input time series [N x D]}

\item{z}{auxilliary input time series [N x C]}

\item{leadtime}{forecast lead time [scalar >= 0]}

\item{lag_Y}{number of time delays for target [scalar >= 1]}

\item{lag_X}{number of time delays [vector (or scalar >= 1)]}

\item{nval}{number of validation records [scalar]}

\item{ntst}{number of test records [scalar]}

\item{wfm}{wavelet-based forecasting method [string: 'none', 'single', 'within', 'across','single-hybrid','within-hybrid','across-hybrid']}

\item{wt}{type of wavelet transform [string; 'at','modwt']}

\item{wavelet}{scaling filter name [string; 'haar', 'db2', 'db4', 'coif1', etc.]}

\item{decomp_level}{decomposition level [1 < integer << N/2]}

\item{max_decomp_level}{maximum decomposition level [scalar >= decomp_level]}

\item{max_wavelet_length}{maxmimum wavelet filter length [scalar >= length(wavelet)]}

\item{ivsm}{input variable selection (IVS) method
[string:'none','boruta','rrf','ea_cmi_htc','ea_cmi_tol','knn_cmi_tol','knn_cmi_bi_tol',pmis_bic','pcis_bic']}

\item{ivs_param}{parameters for input variable selection method
\itemize{
\item \code{rrf}:
\itemize{
\item ivs_param[1] = ntrees [integer > 1)] usually between 128 and 1000
\item ivs_param[2] = regularization parameter [0 < scalar <= 1]}
\item \code{ea_cmi}:
\itemize{
\item ivs_param[1] =  Hampel test criterion threshold [scalar > 0]}
\item \code{ea_cmi_tol}:
\itemize{
\item ivs_param[1] = CMI/MI threshold [0 < scalar <= 1]
\item ivs_param[2] = no. of nearest neighbours [1 < integer < sample size-1]}
\item \code{knn_cmi_tol}:
\itemize{
\item ivs_param[1] = CMI/MI threshold [0 < scalar <= 1]
\item ivs_param[2] = no. of nearest neighbours [1 < integer < sample size-1]}
}}

\item{ddm}{data-driven model [string: 'spov','knnr','grnn','rrf']}

\item{ddm_param}{data-driven model hyper-parameter(s)
\itemize{
\item 'spov': ddm_param[1] = model order [1 < integer <= 3]
\item 'knnr': ddm_param[1] = no. of nearest neighbours [1 < integer < sample size-1]
\item 'grnn': no ddm_param as kernel bandwidth determined automatically
\item 'rrf': ddm_param[1] = ntrees [(1,500)]
}}

\item{scale_inputs}{scale inputs (predictors) [0,1] [binary]}

\item{scale_target}{scale target (predictand) [0,1] [binary]}

\item{cutoff0}{ensure that all predictions > 0 [binary]}

\item{light}{store all function inputs or neglect 'y', 'x', and 'z' [binary]}

\item{savefile}{save file using function name and current Sys.time() [binary]}
}
\value{
mdl_rslt [list] that contains:
\itemize{
\item mdl [list] that includes all function inputs
(this list does not include 'y', 'x', and 'z' if 'light' = \code{TRUE})
\item rslt [list],
\itemize{
\item predictions,
\item target,
\item residual,
\item n.inds,
\item indc,
\item indv,
\item indt,
\item ivs,
\item ddm,
\item perf_c,
\item perf_v,
\item perf_t}
}
}
\description{
This function calibrates a model and generates predictions using the
Wavelet Data-Driven Forecasting Framework (WDDFF) based on the work in
Quilty and Adamowski (2018).
}
\references{
Quilty, J., J. Adamowski, B. Khalil, and M. Rathinasamy (2016), Bootstrap rank-
ordered conditional mutual information (broCMI): A nonlinear input variable
selection method for water resources modeling, Water Resour. Res., 52,
doi:10.1002/2015WR016959.

Quilty, J., and J. Adamowski (2018), Addressing the incorrect usage of wavelet-
based hydrological and water resources forecasting models for real-world
applications with best practices and a new forecasting framework, J. Hydrol.,
doi:10.1016/j.jhydrol.2018.05.003.
}
